LOTS!

+++ CardEditor.xaml.cs
@@ -1,24 +1,14 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
 using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
 using System.Windows.Input;
 using System.Windows.Media;
 using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
 using Microsoft.Win32;
 using System.IO;
-using System.Windows.Automation;
-using System.Net;
-using Newtonsoft.Json;
 using Card_Creator.Data;
 using System.Diagnostics;
-using System.Runtime.CompilerServices;
-using Microsoft.EntityFrameworkCore.ChangeTracking.Internal;
+using System.Text.RegularExpressions;
 
 namespace Card_Creator {
 
@@ -26,43 +16,90 @@ namespace Card_Creator {
 		MainWindow win = (MainWindow)Application.Current.MainWindow;
 		CardCreatorContext context = new CardCreatorContext();
 		string originalPath;
+		bool isEditingExistingCard = false, selectedType = false;
 
 		public CardEditor() {
 			InitializeComponent();
 			win.SaveCardButton.Click += SaveCardButton_Click;
+			DisplayBackground();
 			GenerateTypes();
 			DisplayTypeButtons();
+		}
+
+		//locks the name if you're editing an existing card
+		public void LockName() {
+			NameText.Text = "Name (Cannot be changed)";
+			NameBoxText.IsReadOnly = true;
+			NameCover.Visibility = Visibility.Visible;
+		}
+
+		//function used to edit existing cards
+		public void PremadeInputs(Card card) {
+			LockName();
+			TypeRules(card.TypeName, false);
+			TypeButtonText.Text = card.TypeName;
+			isEditingExistingCard = true;
+			NameBoxText.Text = card.Name;
+			DescriptionBoxText.Text = card.Description;
+			TypeDisplay.Text = card.TypeName;
+			TypeBorderDisplay.BorderBrush = (SolidColorBrush)(new BrushConverter().ConvertFrom(card.TypeColor));
+			LifeBoxText.Text = card.Life.ToString();
+			DamageBoxText.Text = card.Damage.ToString();
+			ManaBoxText.Text = card.Mana.ToString();
+			originalPath = card.PortraitImagePath;
+			ImageSource pImage = new BitmapImage(new Uri(originalPath));
+			PortraitDisplay.Source = pImage;
+			DisplayBackground();
 		}
 
 		//checks if all the user inputs are valid
 		private bool IsCardValid() {
 			if ((String.IsNullOrWhiteSpace(NameBoxText.Text)) ||
-				(IsOnlyNumbers(NameBoxText.Text)) ||
+				(StartWithNumber(NameBoxText.Text)) ||
 				(String.IsNullOrWhiteSpace(DescriptionBoxText.Text)) ||
-				(IsOnlyNumbers(DescriptionBoxText.Text)) ||
+				(StartWithNumber(DescriptionBoxText.Text)) ||
 				(!int.TryParse(LifeBoxText.Text, out _)) ||
 				(!int.TryParse(DamageBoxText.Text, out _)) ||
-				(!int.TryParse(ManaBoxText.Text, out _))) {
+				(!int.TryParse(ManaBoxText.Text, out _)) || 
+				(PortraitDisplay.Source == null) ||
+				(!selectedType) ||
+				(!VerifyRules())) {
 				return false;
 			}
 			return true;
 		}
 
-		//checks is string only contains numbers
-		private bool IsOnlyNumbers(string text) {
-			foreach (char c in text) {
-				if (c < '0' || c > '9') {
+		//checks if name is already being used
+		private bool IsNameInUse(string name) {
+			MessageBox.Show(name);
+			if (context.Card.Find(name) != null) {
+				if (name == context.Card.Find(name).Name) {
+					return true;
+				} else {
 					return false;
 				}
 			}
-			return true;
+			return false;
+		}
+
+		//checks if string starts with a number
+		private bool StartWithNumber(string text) {
+			char firstChar = text[0];
+			if (int.TryParse(firstChar.ToString(), out _)) {
+				return true;
+			} else {
+				return false;
+			}
 		}
 
 		//displays a popup depending on situation
-		private void DisplayPopup(bool isValid) {
+		private void DisplayPopup(bool isValid, string situ) {
 			if (isValid) {
 				PopupBackground.Visibility = Visibility.Visible;
 				ValidPopup.Visibility = Visibility.Visible;
+			} else if (situ == "BadName") {
+				PopupBackground.Visibility = Visibility.Visible;
+				InvalidNamePopup.Visibility = Visibility.Visible;
 			} else {
 				PopupBackground.Visibility = Visibility.Visible;
 				InvalidPopup.Visibility = Visibility.Visible;
@@ -72,20 +109,34 @@ namespace Card_Creator {
 		//saves card to database and saves image to a folder
 		private void SaveCardButton_Click(object sender, RoutedEventArgs e) {
 			if (!IsCardValid()) {
-				DisplayPopup(false);
+				DisplayPopup(false, "");
+			} else if (IsNameInUse(NameBoxText.Text)) {
+				DisplayPopup(false, "BadName");
 			} else {
 				SaveImage(originalPath);
 				string filename = System.IO.Path.GetFileName(originalPath);
 				string newPath = String.Format(ProcessPath() + "\\cardimages\\" + filename);
 				string backgroundPath = String.Format(ProcessPath() + "\\cardimages\\CardBackground.png");
-				NameTest.Text = newPath;
+				Card newCard = CreateCard(newPath, backgroundPath);
+				if (isEditingExistingCard) {
+					Card oldCard = context.Card.Find(NameBoxText.Text);
+					oldCard.Description = newCard.Description;
+					oldCard.TypeColor = newCard.TypeColor;
+					oldCard.TypeName = newCard.TypeName;
+					oldCard.Life = newCard.Life;
+					oldCard.Damage = newCard.Damage;
+					oldCard.Mana = newCard.Mana;
+					oldCard.PortraitImagePath = newCard.PortraitImagePath;
 
-				context.Card.Add(CreateCard(newPath, backgroundPath));
+				} else {
+					context.Card.Add(newCard);
+				}
 				context.SaveChanges();
-				DisplayPopup(true);
+				DisplayPopup(true, "");
 			}
 		}
 
+		//gets the path into the project
 		private string ProcessPath() {
 			string processPath = System.IO.Path.GetDirectoryName(Process.GetCurrentProcess().MainModule.FileName);
 			return processPath;
@@ -95,18 +146,12 @@ namespace Card_Creator {
 		private Card CreateCard(string portraitImagePath, string backgroundImagePath) {
 			Card card = new Card(
 				NameBoxText.Text, DescriptionBoxText.Text, 
-				"typeName", "typeColor",
+				TypeDisplay.Text, TypeBorderDisplay.BorderBrush.ToString(),
 				Convert.ToInt32(LifeBoxText.Text),
 				Convert.ToInt32(DamageBoxText.Text),
 				Convert.ToInt32(ManaBoxText.Text),
 				portraitImagePath, backgroundImagePath);
 			return card;
-		}
-
-		private string ToBase64(string path) {
-			byte[] imageArray = File.ReadAllBytes(path);
-			string imageBase64 = Convert.ToBase64String(imageArray);
-			return imageBase64;
 		}
 
 		//dynamically creates a custom dropdown menu for types
@@ -120,8 +165,9 @@ namespace Card_Creator {
 			};
 			TypeStackPanel.Children.Add(p);
 			foreach (Type type in types) {
+				string nameEdit = Regex.Replace(type.Name, @"\s+", "");
 				Button b = new Button {
-					Name = type.Name,
+					Name = nameEdit,
 					Content = type.Name,
 					FontSize = 14,
 					FontWeight = FontWeights.Light,
@@ -145,29 +191,6 @@ namespace Card_Creator {
 			}
 		}
 
-		//type selection
-		private void TypeClick(object sender, RoutedEventArgs e) {
-			Button b = sender as Button;
-			TypeRules(b.Name);
-			DisplayTypeButtons();
-		}
-
-		//displays the appropriate rules for the selected type
-		private void TypeRules(string typeName) {
-			var types = context.Type;
-			foreach (Type type in types) {
-				if (type.Name == typeName) {
-					LifeText.Text = "Life (" + type.LifeMin + " - " + type.LifeMax + ")";
-					DamageText.Text = "Damage (" + type.DamageMin + " - " + type.DamageMax + ")";
-					ManaText.Text = "Mana (" + type.ManaMin + " - " + type.ManaMax + ")";
-					LifeBoxText.Text = "";
-					DamageBoxText.Text = "";
-					ManaBoxText.Text = "";
-					break;
-				}
-			}
-		}
-
 		//button for displaying and hiding types
 		public void TypeButtonClick(object sender, EventArgs e) {
 			DisplayTypeButtons();
@@ -184,6 +207,27 @@ namespace Card_Creator {
 			}
 		}
 
+		//type selection
+		private void TypeClick(object sender, RoutedEventArgs e) {
+			Button b = sender as Button;
+			LifeCover.Visibility = Visibility.Hidden;
+			DamageCover.Visibility = Visibility.Hidden;
+			ManaCover.Visibility = Visibility.Hidden;
+			selectedType = true;
+			TypeButtonText.Text = b.Content.ToString();
+			string typeColor = GetTypeColor(b.Content.ToString());
+			TypeDisplay.Text = b.Content.ToString();
+			TypeBorderDisplay.BorderBrush = (SolidColorBrush)(new BrushConverter().ConvertFrom(typeColor));
+			TypeRules(b.Content.ToString(), true);
+			DisplayTypeButtons();
+		}
+
+		//gets the typeColor
+		private string GetTypeColor(string nameType) {
+			Type type = context.Type.Find(nameType);
+			return type.ColorCode;
+		}
+
 		//image button
 		private void ImageSelector_Click(object sender, RoutedEventArgs e) {
 			DisplayImage(SelectImage());
@@ -193,14 +237,11 @@ namespace Card_Creator {
 		private void DisplayImage(string filepath) {
 			if (ImageCheck(filepath) == "Good") {
 				ImageSource imgSource = new BitmapImage(new Uri(filepath));
-				ImageTest.Source = imgSource;
-				//player clicked on good image (no popup needed)
+				PortraitDisplay.Source = imgSource;
 			} else if (ImageCheck(filepath) == "Bad") {
-				//player clicked on bad image (popup)
+				BadImagePopup.Visibility = Visibility.Visible;
 			} else if (ImageCheck(filepath) == "Canceled") {
-				//player click x or cancel (no popup needed)
-			} else {
-				//unknown error (popup?)
+				//dead end
 			}
 		}
 
@@ -211,11 +252,12 @@ namespace Card_Creator {
 			} else {
 				int width, height;
 				using (var stream = new FileStream(filepath, FileMode.Open, FileAccess.Read, FileShare.Read)) {
-					var bitmapFrame = BitmapFrame.Create(stream, BitmapCreateOptions.DelayCreation, BitmapCacheOption.None);
+					var bitmapFrame = BitmapFrame.Create(stream, BitmapCreateOptions.DelayCreation, BitmapCacheOption.OnLoad);
 					width = bitmapFrame.PixelWidth;
 					height = bitmapFrame.PixelHeight;
+					bitmapFrame = null;
 				}
-				if (width == 657 && height == 687) {
+				if (!(width - height > 500) && !(width - height < -30)) {
 					return "Good";
 				} else {
 					return "Bad";
@@ -236,10 +278,13 @@ namespace Card_Creator {
 		}
 
 		//saves the image to a folder inside the project
-		private void SaveImage(string filepath) {
-			string name = System.IO.Path.GetFileName(filepath);
+		public void SaveImage(string filepath) {
+			string dirPath = ProcessPath() + "\\cardimages\\";
+			string name = Path.GetFileName(filepath);
 			string destinationPath = GetDestinationPath(name);
-			File.Copy(filepath, destinationPath, true);
+			if (!DoesImageExist(filepath, dirPath)) {
+				File.Copy(filepath, destinationPath, true);
+			}
 		}
 
 		//gets the new destination path from inside the project
@@ -248,58 +293,144 @@ namespace Card_Creator {
 			return appStartPath;
 		}
 
+		//checks if image already exists inside the directory
+		private bool DoesImageExist(string filepath, string dirPath) {
+			string name = System.IO.Path.GetFileName(filepath);
+			DirectoryInfo directory = new DirectoryInfo(dirPath);
+			FileInfo[] files = directory.GetFiles();
+			foreach (FileInfo file in files) {
+				if (String.Compare(file.Name, name) == 0) {
+					return true;
+				}
+			}
+			return false;
+		}
+
 		//reset button
 		public void ResetButton(object sender, EventArgs e) {
-			Refresh();
+			Refresh(false);
 			RemovePopup();
 		}
 
 		//removes popups
 		private void RemovePopup() {
+			BadImagePopup.Visibility = Visibility.Hidden;
 			PopupBackground.Visibility = Visibility.Hidden;
 			ValidPopup.Visibility = Visibility.Hidden;
 			InvalidPopup.Visibility = Visibility.Hidden;
+			InvalidNamePopup.Visibility = Visibility.Hidden;
 		}
 
-		//defaults all the input values
-		public void Refresh() {
-			RemovePopup();
-			NameBoxText.Text = "";
-			DescriptionBoxText.Text = "";
+		//displays the appropriate rules for the selected type
+		private void TypeRules(string typeName, bool resetInputs) {
+			var types = context.Type;
+			foreach (Type type in types) {
+				if (type.Name == typeName) {
+					LifeText.Text = "Life (" + type.LifeMin + " - " + type.LifeMax + ")";
+					DamageText.Text = "Damage (" + type.DamageMin + " - " + type.DamageMax + ")";
+					ManaText.Text = "Mana (" + type.ManaMin + " - " + type.ManaMax + ")";
+					if (resetInputs) {
+						LifeBoxText.Text = "";
+						DamageBoxText.Text = "";
+						ManaBoxText.Text = "";
+					}
+					break;
+				}
+			}
+		}
+
+		//resets the displayed rules
+		private void ResetRules() {
+			LifeText.Text = "Life ()";
+			DamageText.Text = "Damage ()";
+			ManaText.Text = "Mana ()";
 			LifeBoxText.Text = "";
 			DamageBoxText.Text = "";
 			ManaBoxText.Text = "";
 		}
 
-		private void PreviewButton_Click(object sender, RoutedEventArgs e) {
-			//send user to preview with the finished card object
-			//win.DisplayCard(card, db);
-			//win.NavigatePreview();
+		//verifies that the rules are being followed
+		private bool VerifyRules() {
+			Type type = context.Type.Find(TypeDisplay.Text);
+			if ((CompareNumbers(type.LifeMin.ToString(), LifeBoxText.Text)) &&
+				(CompareNumbers(LifeBoxText.Text, type.LifeMax.ToString())) &&
+				(CompareNumbers(type.DamageMin.ToString(), DamageBoxText.Text)) &&
+				(CompareNumbers(DamageBoxText.Text, type.DamageMax.ToString())) &&
+				(CompareNumbers(type.ManaMin.ToString(), ManaBoxText.Text)) &&
+				(CompareNumbers(ManaBoxText.Text, type.ManaMax.ToString()))) {
+				return true;
+			} else {
+				return false;
+			}
 		}
 
+		//compares two numbers
+		private bool CompareNumbers(string first, string second) {
+			int f = Convert.ToInt32(first);
+			int s = Convert.ToInt32(second);
+			if (f <= s) {
+				return true;
+			} else {
+				return false;
+			}
+		}
+
+		//defaults all the input values
+		public void Refresh(bool selected) {
+			selectedType = selected;
+			NameCover.Visibility = Visibility.Hidden;
+			NameBoxText.IsReadOnly = false;
+			NameText.Text = "Name";
+			TypeStackPanel.Children.Clear();
+			GenerateTypes();
+			RemovePopup();
+			ResetRules();
+			isEditingExistingCard = false;
+			originalPath = "";
+			TypeBorderDisplay.BorderBrush = Brushes.Transparent;
+			TypeDisplay.Text = "";
+			PortraitDisplay.Source = null;
+			NameBoxText.Text = "";
+			DescriptionBoxText.Text = "";
+			if (!selected) {
+				TypeButtonText.Text = "Type";
+				LifeCover.Visibility = Visibility.Visible;
+				DamageCover.Visibility = Visibility.Visible;
+				ManaCover.Visibility = Visibility.Visible;
+			} else {
+				LifeCover.Visibility = Visibility.Hidden;
+				DamageCover.Visibility = Visibility.Hidden;
+				ManaCover.Visibility = Visibility.Hidden;
+			}
+		}
+
+		//navigates to the preview page with the current card
+		private void PreviewButton_Click(object sender, RoutedEventArgs e) {
+			win.NavigatePreviewFromDatabase(context.Card.Find(NameDisplay.Text));
+		}
+
+		//displays the background image on the card displayer
+		private void DisplayBackground() {
+			string backgroundPath = String.Format(ProcessPath() + "\\cardimages\\CardBackground.png");
+			ImageSource bgImage = new BitmapImage(new Uri(backgroundPath));
+			BackgroundDisplay.Source = bgImage;
+		}
+
+		//navigates to the collection page
 		private void CollectionButton_Click(object sender, RoutedEventArgs e) { win.NavigateCollection(); }
 
-		private void RestartButton_Click(object sender, RoutedEventArgs e) { Refresh(); }
+		//restart button
+		private void RestartButton_Click(object sender, RoutedEventArgs e) { Refresh(false); }
 
+		//ok button
 		private void OKButton_Click(object sender, RoutedEventArgs e) { RemovePopup(); }
 
-		private void NameBoxText_TextChanged(object sender, TextChangedEventArgs e) {
-			//testName.Text = name; 
-		}
-
 		//updates the displayed card info when input is changed
-		private void DescriptionBoxText_TextChanged(object sender, TextChangedEventArgs e) {
-			//DescriptionTest.Text = DescriptionBoxText.Text;
-		}
-		private void LifeBoxText_TextChanged(object sender, TextChangedEventArgs e) {
-			//LifeTest.Text = LifeBoxText.Text;
-		}
-		private void DamageBoxText_TextChanged(object sender, TextChangedEventArgs e) {
-			//DamageTest.Text = DamageBoxText.Text;
-		}
-		private void ManaBoxText_TextChanged(object sender, TextChangedEventArgs e) {
-			//ManaTest.Text = ManaBoxText.Text;
-		}
+		private void NameBoxText_TextChanged(object sender, TextChangedEventArgs e) { NameDisplay.Text = NameBoxText.Text; }
+		private void DescriptionBoxText_TextChanged(object sender, TextChangedEventArgs e) { DescriptionDisplay.Text = DescriptionBoxText.Text; }
+		private void LifeBoxText_TextChanged(object sender, TextChangedEventArgs e) { LifeDisplay.Text = LifeBoxText.Text; }
+		private void DamageBoxText_TextChanged(object sender, TextChangedEventArgs e) { DamageDisplay.Text = DamageBoxText.Text; }
+		private void ManaBoxText_TextChanged(object sender, TextChangedEventArgs e) { ManaDisplay.Text = ManaBoxText.Text; }
 
 		//this is used for the custom border/textbox combination on the page
 		private void TextBox_LostFocus(object sender, RoutedEventArgs e) { NameBox.BorderThickness = new Thickness(0); }
